Removed backwards compatibility with 2.2 to fix callbacks
